requirements.txt
matplotlib==3.5.1
mysql_connector_repackaged==0.3.1
networkx==2.8.8
numpy==1.23.1
pandas==1.5.1
scikit_learn==1.1.3
stellargraph==1.2.1
tensorflow==2.10.0
